backend/requirements.txt
Flask
Flask-CORS
Flask-PyMongo
python-dotenv
scikit-learn
pandas
numpy
twilio